Create mouseFPV preset for FPVCycle Sonicare
The FPVCycle Sonicare is a new frame from FPVCycle. This frame has a unique wishbone arm design that uses 8 individual arm members to build the 4 arms. This quad has amazing resonance performance and very little low freq noise, meaning we can push the pids pretty far and reduce filtering significantly.

This tune loads the default tune and filters, then applies a non-rpm filtering setup as well as proper pids. There are options for ds300/600 RPM filtering setups that re-load the default filters, apply the proper dshot and filter settings. dshot300 assumes a 4k pidloop will be needed, and thus adds in an lpf2 on the gyro.

There is a dynamic idle option that does not enable dshot, so if things are not properly setup, it will set the value but it wont actually do anything.