mouseFPV TP3 Preset updates for 4.3 and 4.4
This preset has been out for a while, and the only feedback I have received was that some users needed to back off their master just a hair, though this was seemingly rare. This lowers the master a touch to make this work for the most people. Also cleaned up some wording that no longer makes since with the fall of the mpu6000 as our primary gyro, and cleaned up a bad comment.

This also updated the preset to COMMUNITY